folder refactor
refactor code into src and tests folders as decided by the group, fixes #40
